Temporarily disable arbitrary ticket formatting
Currently, a user can enter arbitrary strings, which means that they can make
Underscore say anything, including stuff that activates other bots. This is a
rather easy way to troll, and the feature is not really used at all, so I might
as well disable it until I find an alternative method.