Add redmine.py and shadow/* to .gitignore
redmine.py was something I found browsing around information on using Redmine's
REST API with python. It's not developed by me, so I didn't want to include it,
and I intend to replace it with something simpler that I wrote.

The folder shadow/ is intended to be a directory to place files that I want git
to ignore without having to manually add each to .gitignore.