Change callback method to flat list
Every callback will be called with every event received, and it is up to each
callback to determine whether the event is relevant to it. This allows for
callbacks that can respond to multiple types of events. For example, a whois
callback needs to respond to several different types of methods, including a 318
which signifies the end of the WHOIS reply, and 401 which signifies "no valid
user"